// hub/frontend/lib/api/generated/model/registryPublishedServerListResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Wardn Hub API
 * OpenAPI spec version: 0.2.131
 */
import type { RegistryCatalogServerRead } from './registryCatalogServerRead';
import type { RegistryPageMetadata } from './registryPageMetadata';

export interface RegistryPublishedServerListResponse {
  metadata: RegistryPageMetadata;
  servers: RegistryCatalogServerRead[];
}