// hub/frontend/lib/api/generated/model/registryServerListResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Wardn Hub API
 * OpenAPI spec version: 0.2.131
 */
import type { RegistryListMetadata } from './registryListMetadata';
import type { RegistryServerRead } from './registryServerRead';

export interface RegistryServerListResponse {
  metadata: RegistryListMetadata;
  servers: RegistryServerRead[];
}